Change exponent to 2.75 for powerlaw with 2 floated params
Powerlaw with 3 floated params has bad fit quality (HESSE warnings
from Minuit) but still has better match; e.g. smaller χ² between
fitted pdf and dataset when compared with fit with 2 parameters
floated.

    Floating Parameter    FinalValue +/-  Error
  --------------------  --------------------------
              exponent    2.7576e+00 +/-  7.85e-02
                offset   -7.2659e-02 +/-  1.08e-02
                turnon    1.7071e+03 +/-  1.65e+01

The error for offset is ~14% but the exponent is low, so the idea is
to fix the exponent to the above value and do a fit with 2 floated
parameters.