Implement acceptance ratio correction
PowLawAcceptance now accepts an extra RooAbsReal argument that acts as
a ratio correction to the acceptance.  This can be anything: a
constant, a function or a PDF.

When reading older version of the class ROOT will complain about
unexpected bytes, but those can be ignored; the class default
constructor ensures the instance stays consistent by setting the
correction to unity.

Example use (in python):

  # hcorr is a 1-D histogram
  ardhist = RooDataHist('ardhist', 'DsK/DsPi acceptance ratio datahist',
                        RooArgList(time), RooFit.Import(hcorr, False))
  accratio = RooHistFunc('accratio', 'DsK/DsPi acceptance ratio',
                         RooArgSet(time), ardhist)

  # (1) creating a new acceptance function
  cacceptance = PowLawAcceptance('cacceptance', 'Corrected Power law acceptance',
                                 turnon, time, offset, exponent, beta, accratio)

  # (2) resetting the correction of an existing acceptance function
  cacceptance = PowLawAcceptance(acceptance, 'cacceptance', accratio)