Shift time by average time (t → t-t₀)
This is supposed to reduce parameter correlations, but I'm not
conviced it works when you have a shape that is symmetric in the
observable, e.g. straight line.  Manuel cited track fits, I agree it
works there but that's because they have 2 observables.

Update: showed no improvement to linear ratio fit with this change