tmva/skim: disable a cleaning cut common to both
Not clear if the variable used is correct.

We want χ² of the vertex formed by the bachelor track and D-children
tracks to be large.  This ensures we have a particle with a lifetime
after the B decays.

The χ² of the 3 D-children tracks however should be small since they
form a nice vertex (they are coming from the same D).  The variable
"lab2_ENDVERTEX_CHI2" might be the latter, need to resolve this
confusion.

TODO: look at the distribution for all events (sig + bkg)