tmva: force all corrn plots to same colour scale
ROOT chooses the rainbow colour map by default (good), but the colour
stops are not easily controlled.  It would be nice to also change the
stops such that we can highlight different ranges with different
"family" of colours; so something like: -20–20 green, -50–-20 cyan,
-85–-50 blue, -85> violet, similarly 20–50 yellow, 50–85 orange, >85
red.  AFAIU, this can be done by setting the colour stops in
proportional intervals.