tmva: th1signedsum sums over bin content
This is akin to making a sum over the bin contents, mind you, it is not
to be confused with an integral.  It maybe thought of as integrating
after rotating the histogram by -π/2.

  ∑ᵢcᵢ

The idea is to have equivalent to the above, but the granularity is lost
during binning.  It is assumed the bin centre is a faithful
representation.  This is definitely not true in the case of variable bin
widths, or a rapidly changing distribution, but can be considered a
reasonable compromise.