tacc: major bug fix, weighted datasets
Dsπ and DsK datasets have different event weights.  This is fine, but
the datasets need to be combined using categories for a simultaneous
fit.  RooFit however only uses the first weight variable name after the
append, which means event weights for the second dataset are
wrong/ignored (don't know which).  This is a RooFit bug in my opinion.
To avoid the problem, fill the dataset by hand, with same weight
variable name, circumventing the problem.

TODO: File bug report against RooFit.